back-end/account/requirements.txt
fastapi
SQLAlchemy
uvicorn
httpx
psycopg2
PyYAML
python-jose
pycryptodome
python-dotenv
python-multipart
